created struct for storing vertices and indices

+++ src/buffers/VertexAttributes.cpp
@@ -7,7 +7,7 @@
 
 void VertexAttributes::add(int size,int stride,int offset)
 {
-    glVertexAttribPointer(attributes, size, GL_FLOAT, GL_FALSE, stride * sizeof(int), (void*)(offset * sizeof(int)));
+    glVertexAttribPointer(attributes, size, GL_FLOAT, GL_FALSE, stride * sizeof(float), (void*)(offset * sizeof(float)));
     glEnableVertexAttribArray(attributes);
     attributes++;
 }
